Do not limit datastream write to uint32_t size - 2.0

+++ include/fc/io/datastream.hpp
@@ -37,7 +37,7 @@ class datastream {
       }
 
       inline bool write( const char* d, size_t s ) {
-        if( _end - _pos >= (int32_t)s ) {
+        if( size_t(_end - _pos) >= (size_t)s ) {
           memcpy( _pos, d, s );
           _pos += s;
           return true;
